Refactor: 수익률 구하는 클래스 분리
