+++ src/main/java/lotto/view/LottoOutputView.java
@@ -1,23 +1,25 @@
 package lotto.view;
 
 import lotto.domain.LottoProfitRate;
-import lotto.domain.LottoResult;
+import lotto.domain.LottoStatistics;
+import lotto.domain.LottoTotalWinnings;
 import lotto.model.Lotto;
 import lotto.model.LottoBuyer;
+import lotto.model.LottoResults;
 import lotto.model.LottoWinningNumber;
-
-import java.util.List;
 
 public class LottoOutputView {
     private static String CONFIRM_BUY_AMOUNT_MESSAGE = "%d개를 구매했습니다.\n";
     private static String STATISTIC_MESSAGE = "당첨 통계\n---";
     private static String PROFIT_RATE_MESSAGE = "총 수익률은 %.1f%%입니다.\n";
     LottoProfitRate lottoProfitRate = new LottoProfitRate();
-    LottoResult lottoResult = new LottoResult();
+    LottoStatistics lottoStatistics = new LottoStatistics();
+    LottoTotalWinnings lottoTotalWinnings = new LottoTotalWinnings();
 
     public void printLottoStatistics(LottoBuyer lottoBuyer, LottoWinningNumber lottoWinningNumber) {
         System.out.println(STATISTIC_MESSAGE);
-        lottoResult.getLottoStatistics(lottoBuyer, lottoWinningNumber);
+        lottoStatistics.getLottoStatistics(lottoBuyer, lottoWinningNumber);
+        printRankingResult(lottoBuyer.getLottoResults());
     }
 
     public void printBuyAmount(LottoBuyer lottoBuyer) {
@@ -31,7 +33,17 @@ public class LottoOutputView {
     }
 
 
+    private void printRankingResult(LottoResults lottoResults) {
+        System.out.println("3개 일치 (5,000원) - " + lottoResults.getLottoPrizeCount("FIFTH_PLACE") + "개");
+        System.out.println("4개 일치 (50,000원) - " + lottoResults.getLottoPrizeCount("FOURTH_PLACE") + "개");
+        System.out.println("5개 일치 (1,500,000원) - " + lottoResults.getLottoPrizeCount("THIRD_PLACE") + "개");
+        System.out.println("5개 일치, 보너스 볼 일치 (30,000,000원) - " + lottoResults.getLottoPrizeCount("SECOND_PLACE") + "개");
+        System.out.println("6개 일치 (2,000,000,000원) - " + lottoResults.getLottoPrizeCount("FIRST_PLACE") + "개");
+    }
+
+
     public void printLottoProfitRate(LottoBuyer lottoBuyer) {
+        lottoTotalWinnings.getTotalWinnings(lottoBuyer);
         double profitRate = lottoProfitRate.getProfitRate(lottoBuyer.getTotalWinnings(), lottoBuyer.getPayment());
         System.out.printf(PROFIT_RATE_MESSAGE, profitRate);
     }
